Updated computePath prototype

+++ src/Vehicle.cpp
@@ -1,5 +1,5 @@
 #include "Vehicle.hpp"
 
-Vehicle::Vehicle(int x, int y, int s, int d, int yaw, int speed): 
+Vehicle::Vehicle(double x, double y, double s, double d, double yaw, double speed): 
                       x_(x), y_(y), s_(s), d_(d), yaw_(yaw), speed_(speed) { 
 }
